sr_tg/requirements.txt
accelerate==0.24.1
bitsandbytes==0.40.2
transformers==4.34.0
scipy==1.11.4
scikit-learn==1.3.2
peft==0.5.0
fastapi==0.95.0
uvicorn==0.21.1
python-multipart
openai-whisper
torch